tests/wpt/tests/css/css-text/parsing/text-align-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: getComputedStyle().textAlign</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-align">
<meta name="assert" content="text-align computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #container {
    text-align: center;
  }
</style>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
test_computed_value("text-align", "start");
test_computed_value("text-align", "end");
test_computed_value("text-align", "left");
test_computed_value("text-align", "right");
test_computed_value("text-align", "center");
test_computed_value("text-align", "justify");
test_computed_value("text-align", "match-parent", "center");

// TODO: Test justify-all
</script>
</body>
</html>
